requirements.txt
Faker==37.12.0
kafka-python==2.2.16
psycopg2-binary==2.9.11
python-dotenv==1.2.1
tzdata==2025.2
